# Conflicts:
#	Sources/Internal/Scene3D/Scene.cpp
#	Sources/Internal/Scene3D/Scene.h


Former-commit-id: 222ef84150fca75cc7b01ba9b7d74207241cad59

+++ Programs/PerfomanceTests/Sources/Tests/Utils/TankUtils.h
@@ -17,8 +17,8 @@ struct TankNode
     const static FastName SKINNED_TANK;
 };
 
-void Animate(Entity* skinnedTank, const Vector<uint16>& jointIndexes, float32 angle);
-void MakeSkinnedTank(Entity* entity, Vector<uint16>& jointsInfo);
+void Animate(Entity* skinnedTank, const Vector<uint32>& jointIndexes, float32 angle);
+void MakeSkinnedTank(Entity* entity, Vector<uint32>& jointsInfo);
 }
 }
